+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/impl/OsUtils.kt
@@ -21,8 +21,11 @@ internal val currentOS: OS by lazy {
     }
 }
 
+internal val isWindowsOs: Boolean
+    get() = currentOS === OS.Windows
+
 internal fun executableName(nameWithoutExtension: String): String =
-    if (currentOS == OS.Windows) "$nameWithoutExtension.exe" else nameWithoutExtension
+    if (isWindowsOs) "$nameWithoutExtension.exe" else nameWithoutExtension
 
 internal fun javaExecutable(javaHome: String): String =
     File(javaHome).resolve("bin/${executableName("java")}").absolutePath