+++ fluxo-kmp-conf/src/main/kotlin/fluxo/gradle/CliArgUtils.kt
@@ -1,5 +1,6 @@
 package fluxo.gradle
 
+import fluxo.conf.impl.isWindowsOs
 import java.io.File
 import org.gradle.api.file.FileSystemLocation
 import org.gradle.api.provider.Provider
@@ -14,7 +15,7 @@ internal fun <T : Any?> MutableCollection<String>.cliArg(
         if (value) add(name)
     } else if (value != null) {
         add(name)
-        add(fn(value))
+        add(cliEscaped(fn(value)))
     }
 }
 
@@ -27,9 +28,8 @@ internal fun <T : Any?> MutableCollection<String>.cliArg(
     cliArg(name = name, value = value.orNull, base = base, fn = fn)
 }
 
-internal fun MutableCollection<String>.javaOption(value: String) {
-    cliArg(name = "--java-options", value = "'$value'")
-}
+internal fun MutableCollection<String>.javaOption(value: String) =
+    cliArg(name = "--java-options", value = value)
 
 private fun <T : Any?> defaultToString(base: File? = null): (T) -> String =
     {
@@ -38,5 +38,27 @@ private fun <T : Any?> defaultToString(base: File? = null): (T) -> String =
             is File -> it.normalizedPath(base)
             else -> it.toString()
         }
-        "\"$asString\""
+        cliEscaped(asString)
     }
+
+/**
+ * Return CLI argument string for the given [value].
+ * If the value contains spaces or special characters, returns quoted and escaped.
+ * On Windows, it uses double quotes, single quotes on other platforms.
+ */
+private fun cliEscaped(value: String): String {
+    if (!value.requiresQuotes()) {
+        return value
+    }
+    val q = if (isWindowsOs) '"' else '\''
+    return "$q${value.replace("$q", "\\$q")}$q"
+}
+
+private fun String.requiresQuotes(): Boolean {
+    return isEmpty() || any {
+        it.isWhitespace() || when (it) {
+            '*', '?', '[', ']', '(', ')', '$', '|', ';', '&', '"', '\'' -> true
+            else -> false
+        }
+    }
+}